fixed bug with -inf when casting down to float16

+++ PoroX/models/defaults.py
@@ -1,3 +1,3 @@
 import jax.numpy as jnp
 
-DEFAULT_DTYPE = jnp.float32
+DEFAULT_DTYPE = jnp.float16
